update flash px4 script

+++ scripts/flash_px4.sh
@@ -12,7 +12,16 @@ fi
 
 echo "Flashing firmware: $FW_PATH"
 
+SERIALDEVICE=$(ls -l /dev/serial/by-id/*ARK* | awk -F'/' '{print "/dev/"$NF}')
+
+# if it fails to find the ARKV6X quit the script
+if [ $? -ne 0 ]; then
+    echo "ARKV6X not found, quitting"
+    exit 1
+fi
+echo $SERIALDEVICE
+
 sudo systemctl stop mavlink-router
 python3 /usr/bin/reset_fmu_wait_bl.py
-python3 /usr/bin/px_uploader.py --port /dev/ttyACM0 $FW_PATH
+python3 /usr/bin/px_uploader.py --port $SERIALDEVICE $FW_PATH
 sudo systemctl start mavlink-router
